Refactor volumes reducer
Remove `FormReducer` as the form uses the specific
typed reducers like external volumes, host volumes
and local volumes.

close DCOS-14727
